Add files via upload
As part of learning to code, I completed a project that involved creating a basic Spotify clone using HTML, CSS, and JavaScript. This project was designed to showcase my skills in front-end web development. It is a responsive and user-friendly web application that allows users to play music. I used HTML and CSS to design and style the interface, and I implemented JavaScript to handle user input and interaction such as basic playback controls - play, pause, and next. I also gained experience working with version control tools such as Git and GitHub.